VS builds: - Added simcqt11.vxproj: Builds gui with VS11 and Qt5.0.2 - Tried to resolve problems with wrong libraries for the simcqt.vxproj ( VS10 / Qt4.8.4 ) debug package.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16537 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/dbc/dbc.hpp
@@ -420,7 +420,7 @@ public:
   { return ok() ? _tooltip : ""; }
 
   // Helper functions
-  unsigned effect_count() const
+  size_t effect_count() const
   { assert( _effects ); return _effects -> size(); }
 
   bool found() const
